requirements.txt
# 代码知识工程 — 依赖（Python >=3.10）
pydantic>=2.0
pydantic-settings>=2.0
pyyaml>=6.0
javalang>=0.13.0
networkx>=3.0
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
httpx>=0.24.0
streamlit>=1.37.0  # 1.37+ 支持 st.fragment(run_every) 自动刷新进度
# 图数据库与向量库（可选：不安装则 backend 使用 memory）
neo4j>=5.0.0
weaviate-client>=4.0.0
